+++ src/FSharp.Formatting.ApiDocs/GenerateHtml.fs
@@ -513,6 +513,31 @@ type HtmlRender(model: ApiDocModel, ?menuTemplateFolder: string) =
                 div [] (renderMembers "Constructors" "Constructor" constructors)
                 div [] (renderMembers "Instance members" "Instance member" instanceMembers)
                 div [] (renderMembers "Static members" "Static member" staticMembers)
+
+            let inheritedMemberGroups =
+                entity.InheritedMembers
+                |> List.choose (fun (baseTypeHtml, members) ->
+                    let instMembers =
+                        members
+                        |> List.filter (fun m -> m.Kind = ApiDocMemberKind.InstanceMember && not m.IsObsolete)
+
+                    let statMembers =
+                        members
+                        |> List.filter (fun m -> m.Kind = ApiDocMemberKind.StaticMember && not m.IsObsolete)
+
+                    if not (List.isEmpty instMembers) || not (List.isEmpty statMembers) then
+                        Some(baseTypeHtml, instMembers, statMembers)
+                    else
+                        None
+                )
+
+            if not (List.isEmpty inheritedMemberGroups) then
+                h3 [] [ !!"Inherited members" ]
+
+                for (baseTypeHtml, instMembers, statMembers) in inheritedMemberGroups do
+                    h4 [] [ !!"Inherited from "; embed baseTypeHtml ]
+                    div [] (renderMembers "Instance members" "Instance member" instMembers)
+                    div [] (renderMembers "Static members" "Static member" statMembers)
         ]
 
     let namespaceContent (nsIndex, ns: ApiDocNamespace) =